+++ libs/ng-utils/css-props/css-props.service.ts
@@ -27,7 +27,7 @@ export class CssPropsService {
 
   public updateElementVars(element: HTMLElement, cssProps: any) {
     this.ngZone.runOutsideAngular(() => {
-      updateCssProps(this.document.body, cssProps);
+      updateCssProps(element, cssProps);
     });
   }
 }