+++ libs/ng-utils/css-props/public_api.ts
@@ -3,3 +3,4 @@ export * from './css-props.directive';
 export * from './css-props.service';
 export * from './update-css-props';
 export * from './update-dynamic-style';
+export * from './element-css.service';