+++ libs/ng-utils/morphdom/morphdom.directive.spec.ts
@@ -1,0 +1,8 @@
+import { MorphdomDirective } from './morphdom.directive';
+
+describe('MorphdomDirective', () => {
+  it('should create an compRef', () => {
+    const directive = new MorphdomDirective();
+    expect(directive).toBeTruthy();
+  });
+});