+++ libs/ng-utils/morphdom/public_api.ts
@@ -1,0 +1,3 @@
+export * from './morphdom.module';
+export * from './morphdom.directive';
+export * from './morphdom.service';